+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 ## 0.2.0 (2023-08-18)
 
 - Updated to newer versions of optional dependencies.
-- Added documentation.
+- Created documentation at readthedocs.
 - Added new text elements `div()` and `span()`. Added new parameters to HTML elements - in particular `cl` and `attrs`.
 - Deprecated `print_html()` function. New `print()` should be used instead.
 - Deprecated `print_code()` function. New `print_source_code()` should be used instead.
 - Replaced code-prettify with highlight.js for code blocks.
 - `print_h1`, ..., `print_h6` functions can now take a `Reference` object.
+- Added `--config-path` CLI option, changed how `pyreball-generate-config` command works and how the config paths are
+  handled.
 
 ## 0.1.1 (2021-09-14)